+++ ts/components/messages/paginated/MessageDetail/MedicalPrescriptionAttachments.tsx
@@ -23,11 +23,6 @@ const styles = StyleSheet.create({
   padded: {
     paddingHorizontal: customVariables.contentPadding
   },
-  image: {
-    width: 300,
-    height: BARCODE_HEIGHT,
-    resizeMode: "contain"
-  },
   note: {
     lineHeight: 16,
     paddingTop: 4