+++ ts/features/bonus/bpd/screens/optInPaymentMethods/OptInPaymentMethodsThankYouDeleteMethodsScreen.tsx
@@ -1,7 +1,9 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
+import { useNavigation } from "@react-navigation/native";
 import React, { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import I18n from "../../../../../i18n";
+import ROUTES from "../../../../../navigation/routes";
 import { useIOSelector } from "../../../../../store/hooks";
 import { deleteAllPaymentMethodsByFunctionSelector } from "../../../../../store/reducers/wallet/wallets";
 import { showToast } from "../../../../../utils/showToast";
@@ -14,14 +16,12 @@ import {
   isReady,
   isUndefined
 } from "../../model/RemoteValue";
-import {
-  optInPaymentMethodsCompleted,
-  optInPaymentMethodsDeletionChoice
-} from "../../store/actions/optInPaymentMethods";
+import { optInPaymentMethodsDeletionChoice } from "../../store/actions/optInPaymentMethods";
 import { optInStatusSelector } from "../../store/reducers/details/activation";
 
 const OptInPaymentMethodsThankYouDeleteMethodsScreen = () => {
   const dispatch = useDispatch();
+  const navigation = useNavigation();
   const deleteAllPaymentMethodsByFunctionStatus = useIOSelector(
     deleteAllPaymentMethodsByFunctionSelector
   );
@@ -39,9 +39,14 @@ const OptInPaymentMethodsThankYouDeleteMethodsScreen = () => {
       isReady(deleteAllPaymentMethodsByFunctionStatus)
     ) {
       showToast(I18n.t("bonus.bpd.optInPaymentMethods.thankYouPage.toast"));
-      dispatch(optInPaymentMethodsCompleted());
+      navigation.navigate(ROUTES.WALLET_HOME);
     }
-  }, [optInStatus, deleteAllPaymentMethodsByFunctionStatus, dispatch]);
+  }, [
+    optInStatus,
+    deleteAllPaymentMethodsByFunctionStatus,
+    dispatch,
+    navigation
+  ]);
 
   const isOptInStatusLoading = pot.fold(
     optInStatus,