+++ ts/navigation/AppStackNavigator.tsx
@@ -15,7 +15,13 @@ import { useRef } from "react";
 import { IOColors } from "../components/core/variables/IOColors";
 import workunitGenericFailure from "../components/error/WorkunitGenericFailure";
 import LoadingSpinnerOverlay from "../components/LoadingSpinnerOverlay";
-import { bpdEnabled, fimsEnabled, myPortalEnabled, svEnabled } from "../config";
+import {
+  bpdEnabled,
+  bpdOptInPaymentMethodsEnabled,
+  fimsEnabled,
+  myPortalEnabled,
+  svEnabled
+} from "../config";
 import BPD_ROUTES from "../features/bonus/bpd/navigation/routes";
 import { CdcStackNavigator } from "../features/bonus/cdc/navigation/CdcStackNavigator";
 import { CDC_ROUTES } from "../features/bonus/cdc/navigation/routes";
@@ -32,6 +38,10 @@ import {
   FimsNavigator
 } from "../features/fims/navigation/navigator";
 import FIMS_ROUTES from "../features/fims/navigation/routes";
+import {
+  IDPayOnboardingNavigator,
+  IDPayOnboardingRoutes
+} from "../features/idpay/onboarding/navigation/navigator";
 import UADONATION_ROUTES from "../features/uaDonations/navigation/routes";
 import { UAWebViewScreen } from "../features/uaDonations/screens/UAWebViewScreen";
 import { ZendeskStackNavigator } from "../features/zendesk/navigation/navigator";
@@ -45,6 +55,7 @@ import {
   sessionTokenSelector
 } from "../store/reducers/authentication";
 import {
+  bpdRemoteConfigSelector,
   isCdcEnabledSelector,
   isCGNEnabledSelector,
   isFIMSEnabledSelector
@@ -172,6 +183,11 @@ export const AppStackNavigator = (): React.ReactElement => {
           component={CdcStackNavigator}
         />
       )}
+
+      <Stack.Screen
+        name={IDPayOnboardingRoutes.IDPAY_ONBOARDING_MAIN}
+        component={IDPayOnboardingNavigator}
+      />
     </Stack.Navigator>
   );
 };
@@ -190,6 +206,10 @@ const InnerNavigationContainer = (props: { children: React.ReactElement }) => {
 
   const cgnEnabled = useIOSelector(isCGNEnabledSelector);
   const isFimsEnabled = useIOSelector(isFIMSEnabledSelector) && fimsEnabled;
+
+  const bpdRemoteConfig = useIOSelector(bpdRemoteConfigSelector);
+  const isOptInPaymentMethodsEnabled =
+    bpdRemoteConfig?.opt_in_payment_methods_v2 && bpdOptInPaymentMethodsEnabled;
 
   const linking: LinkingOptions = {
     enabled: true,
@@ -220,9 +240,17 @@ const InnerNavigationContainer = (props: { children: React.ReactElement }) => {
             [ROUTES.PAYMENTS_HISTORY_SCREEN]: "payments-history",
             [ROUTES.CREDIT_CARD_ONBOARDING_ATTEMPTS_SCREEN]:
               "card-onboarding-attempts",
-            ...(bpdEnabled
-              ? { [BPD_ROUTES.CTA_BPD_IBAN_EDIT]: "bpd-iban-update" }
-              : {})
+            ...(bpdEnabled && {
+              [BPD_ROUTES.CTA_BPD_IBAN_EDIT]: "bpd-iban-update"
+            }),
+            ...(isOptInPaymentMethodsEnabled && {
+              [BPD_ROUTES.OPT_IN_PAYMENT_METHODS.MAIN]: {
+                path: "bpd-opt-in",
+                screens: {
+                  [BPD_ROUTES.OPT_IN_PAYMENT_METHODS.CHOICE]: "choice"
+                }
+              }
+            })
           }
         },
         [ROUTES.SERVICES_NAVIGATOR]: {
@@ -234,17 +262,18 @@ const InnerNavigationContainer = (props: { children: React.ReactElement }) => {
                 activate: activate => activate === "true"
               }
             },
-            ...(myPortalEnabled ? { [ROUTES.SERVICE_WEBVIEW]: "webview" } : {}),
-            ...(svEnabled ? svLinkingOptions : {})
+            ...(myPortalEnabled && { [ROUTES.SERVICE_WEBVIEW]: "webview" }),
+            ...(svEnabled && svLinkingOptions)
           }
         },
-        ...(isFimsEnabled ? fimsLinkingOptions : {}),
-        ...(cgnEnabled ? cgnLinkingOptions : {}),
+        ...(isFimsEnabled && fimsLinkingOptions),
+        ...(cgnEnabled && cgnLinkingOptions),
         [UADONATION_ROUTES.WEBVIEW]: "uadonations-webview",
         [ROUTES.WORKUNIT_GENERIC_FAILURE]: "*"
       }
     }
   };
+
   return (
     <NavigationContainer
       theme={IOTheme}