+++ ts/navigation/params/OnboardingParamsList.ts
@@ -1,4 +1,5 @@
 import { FingerprintScreenNavigationParams } from "../../screens/onboarding/FingerprintScreen";
+import { OnboardingNotificationsPreferencesScreenNavigationParams } from "../../screens/onboarding/OnboardingNotificationsPreferencesScreen";
 import { OnboardingServicesPreferenceScreenNavigationParams } from "../../screens/onboarding/OnboardingServicesPreferenceScreen";
 import ROUTES from "../routes";
 
@@ -12,4 +13,5 @@ export type OnboardingParamsList = {
   [ROUTES.INSERT_EMAIL_SCREEN]: undefined;
   [ROUTES.READ_EMAIL_SCREEN]: undefined;
   [ROUTES.ONBOARDING_COMPLETED]: undefined;
+  [ROUTES.ONBOARDING_NOTIFICATIONS_PREFERENCES]: OnboardingNotificationsPreferencesScreenNavigationParams;
 };